Move asset copying logic to pkg
